Shopping cart update(Price)
