+++ ConstructionLK/ConstructionLK.csproj
@@ -287,6 +287,7 @@
     <Compile Include="Controllers\ItemsController.cs" />
     <Compile Include="Controllers\LocationsController.cs" />
     <Compile Include="Controllers\ManageController.cs" />
+    <Compile Include="Controllers\NewOrderController.cs" />
     <Compile Include="Controllers\OrderController.cs" />
     <Compile Include="Controllers\ProductController.cs" />
     <Compile Include="Controllers\PublishedItemsController.cs" />
@@ -891,6 +892,8 @@
     <Content Include="Views\PublishedItems\Edit.cshtml" />
     <Content Include="Views\PublishedItems\Index.cshtml" />
     <Content Include="Views\Items\BuyProduct.cshtml" />
+    <Content Include="Views\Orders\cartold.cshtml" />
+    <Content Include="Views\NewOrder\cart.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
