+++ ConstructionLK/Views/Items/BuyProduct.cshtml
@@ -7,7 +7,8 @@
 
 <div class="panel panel-success">
     <div class="panel-heading">
-        <h3 class="panel-title">@Html.DisplayFor(model => model.ItemName)</h3>
+        @*<h3 class="panel-title">@Html.DisplayFor(model => model.ItemName)</h3>*@
+        <h3 class="panel-title">@Html.Raw(ViewBag.idata.ItemName)</h3>
     </div>
     <div class="panel-body">
 
@@ -15,27 +16,37 @@
             @*<tr>
                 <td>Product Name</td>
                 <td style="padding-left:70px;">@Html.DisplayFor(model => model.ItemName)</td>
+                <td style="padding-left:70px;">@Html.Raw(ViewBag.ItemName)</td>
             </tr>*@
             <tr>
                 <td>Product code</td>
-                <td style="padding-left:70px;">@Html.DisplayFor(model => model.ItemCode)</td>
+                @*<td style="padding-left:70px;">@Html.DisplayFor(model => model.ItemCode)</td>*@
+                <td style="padding-left:70px;">@Html.Raw(ViewBag.idata.ItemCode)</td>
             </tr>
             <tr>
                 <td>Description</td>
-                <td style="padding-left:70px;">@Html.DisplayFor(model => model.Description)</td>
+                @*<td style="padding-left:70px;">@Html.DisplayFor(model => model.Description)</td>*@
+                <td style="padding-left:70px;">@Html.Raw(ViewBag.idata.Description)</td>
             </tr>
-            <tr>
+            @*<tr>
                 <td>Status</td>
                 <td style="padding-left:70px;">@Html.DisplayFor(model => model.ItemStatus.Name)</td>
+                <td style="padding-left:70px;">@Html.Raw(ViewBag.idata.ItemStatus)</td>
+            </tr>*@
+            <tr>
+                <td>Unit price</td>
+                <td style="padding-left:70px;">@Html.Raw(ViewBag.data)</td>
+
             </tr>
 
             <tr>
                 <td>Quantity</td>
                 <td style="padding-left:70px;">
-                    @using (Html.BeginForm("order", "Order", FormMethod.Get))
+                    @using (Html.BeginForm("order", "NewOrder", FormMethod.Get))
             {
-                        <input type="hidden" class="form-control" id="id" name="id" value="@Html.DisplayFor(model => model.Id)">
-                        @*<input type="hidden" class="form-control" id="Price" value=100.00 >*@
+                        <input type="hidden" class="form-control" id="id" name="id" value="@Html.Raw(ViewBag.idata.Id)">
+                        <input type="hidden" class="form-control" id="price" name="price" value="@Html.Raw(ViewBag.data)">
+
                         <br/>
                         <input type="text" class="form-control input-sm" id="Qnt" name="Qnt">
                         <br />
