+++ ConstructionLK/Views/Items/ReadOnlyIndex.cshtml
@@ -67,7 +67,7 @@ else
                         @Html.ActionLink("Details", "Details", new { id = item.Id }, new { @class = "btn btn-outline-primary btn-sm" }) |
                         @*@Html.ActionLink("Buy this", "Create","ItemRequests", new { id = item.Id }, new { @class = "btn btn-outline-warning btn-sm" })*@
                         @*@Html.ActionLink("Buy this", "Create", "ItemRequests", new { id = item.Id, user = User.Identity.GetUserId(), provider = item.UserId }, new { @class = "btn btn-outline-warning btn-sm" })*@
-                        @Html.ActionLink("Buy this", "Buy", "Items", new {id = item.Id }, new { @class = "btn btn-outline-warning btn-sm"})
+                        @Html.ActionLink("Add to cart", "Buy", "Items", new {id = item.Id }, new { @class = "btn btn-outline-warning btn-sm"})
                         @*@Html.ActionLink("Review and Delete", "Delete", new { id = item.Id }, new { @class = "btn btn-outline-warning btn-sm" })|*@
                         @*<button data-item-id="@item.Id" class="btn btn-outline-warning btn-sm js-delete">Delete Now</button>*@
                     </td>
