+++ ConstructionLK/Views/NewOrder/cart.cshtml
@@ -1,0 +1,72 @@
+﻿@using ConstructionLK.Controllers;
+@using Microsoft.Ajax.Utilities
+@using Microsoft.AspNet.Identity
+@model ConstructionLK.Models.Customer
+@{
+    ViewBag.Title = "cart";
+}
+
+<h2>cart</h2>
+<html>
+<head>
+    <meta name="viewport" content="width=device-width" />
+    <title>ConfirmOrder1</title>
+</head>
+<body>
+
+
+    <table class="table table-bordered table-hover ">
+        <tr class="success">
+            <th>Product</th>
+            <th>unit Price</th>
+            <th>Quantity</th>
+            <th>Sub total</th>
+        </tr>
+
+        @{ 
+                double s = 0;
+                
+            }
+
+            @foreach (Product product in (List<Product>)Session["cart"])
+            {
+
+
+
+           
+            <tr>
+                <td>@product.It.ItemName</td>
+                <td>
+                    @foreach (var price in @product.It.PublishedItems)
+                    {
+                        @price.Price
+                        ViewBag.p = @price.Price;
+                    }
+                </td>
+                <td>@product.Quantity</td>
+
+                <td>@(ViewBag.p*product.Quantity)</td>
+                <td> @Html.ActionLink("Cancel order", "cancelOrder", "NewOrder", new { id = @product.It.Id }, null )</td>
+            </tr>
+                s = s + ViewBag.p * product.Quantity;
+            }
+
+
+        <tr>
+            <td></td>
+            <td></td>
+            <td>Total</td>
+            <td>@s</td>
+        </tr>
+
+    </table>
+
+    <br />
+    @Html.ActionLink("Continue Shopping", "Index", "Items", new { id = 1 }, new { @class = "btn btn-warning" }) @Html.ActionLink("Confirm Order", "", "Items", new { }, new { @class = "btn btn-success" })
+
+</body>
+</html>
+
+
+
+
