+++ ConstructionLK/Views/Orders/cartold.cshtml
@@ -36,7 +36,7 @@
                 {
                     <tr>
                         <td>@product.It.ItemName</td>
-                        <td></td>
+                        <td>@Html.Raw(ViewBag.unitprice)</td>
                         <td>@product.Quantity</td>
                         <td></td>
                         <td> @Html.ActionLink("Cancel order", "cancelOrder", "Order", new { id = @product.It.Id }, new { @class = "btn btn-outline-warning btn-sm" })</td>
